pyproject.toml
[project]
name = "openset-reid-finetune"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "datasets>=3.6.0",
    "fiftyone>=1.5.2",
    "google-generativeai>=0.8.5",
    "huggingface-hub>=0.32.3",
    "numpy>=2.2.6",
    "pillow>=11.2.1",
    "python-dotenv>=1.1.0",
    "torch>=2.7.0",
    "torchvision>=0.22.0",
    "transformers>=4.52.4",
    "wandb>=0.19.11",
]
